Added contrast

+++ processors/imageRotate.py
@@ -5,4 +5,4 @@ import numpy as np
 from utils import helpers as hp
 
 def imageRotate(bmp, degree):
-    return np.rot90(bmp.imageData, k=int(degree/90))
+    return np.rot90(bmp.imageData, k=int(degree/90)).astype(float)
